Bug when using leaveRoom
When using leaveRoom, the complete clientRooms Instance get removed, which causes the SocketIOClient getAllRooms Method to returns a empty list, even when your client is still in some rooms. This doesn't harm SocketIOServer getRoomOperations Method (which still returns all clients from the given room), but harms for example the onDisconnectListener, when you override the onDisconnect Method, which than returns a empty list for the joinedRooms.